requirements.txt
streamlit
transformers
torch
sentence-transformers
openai
python-dotenv